+++ main/src/me/ricky/kbot/core/data/MongoDBHandler.kt
@@ -1,5 +1,11 @@
 package me.ricky.kbot.core.data
 
-class MongoDBHandler {
+import com.mongodb.MongoClient
+import com.mongodb.client.MongoDatabase
+import me.ricky.kbot.core.data.model.MongoConfig
+import org.litote.kmongo.KMongo
 
+class MongoDBHandler(val config: MongoConfig) {
+  val client: MongoClient = KMongo.createClient(config.connection)
+  val database: MongoDatabase = client.getDatabase(config.database)
 }