+++ CHANGELOG.md
@@ -7,3 +7,8 @@
 * Added BotConfig Serializable
 * Added KBot class to make bot instances
 * some other tweaks
+
+#### 0.0.2
+* Added Command Arguments, description, permissions to CommandInfo
+* Added the start of a command dsl
+* Added the start of mongo handler and documents